Relocate directions button to building sheet
This was to avoid the troublesome floating direction button appearing and disappearing when it shouldn't have.
I tried to find a way to keep track of the building sheet's state, but wasn't able to reliably find whether it was on the screen.
It doesn't match the mockup 100% of course, it being in the sheet now, but I think it's better to have that than to have a visible bug at the demo or even a crash.